requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.1.0
langchain-ollama>=0.1.0

# Database and storage
supabase>=2.17.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-optimize>=0.9.0

# Text processing and NLP
beautifulsoup4>=4.12.0
python-Levenshtein>=0.21.0
fuzzywuzzy>=0.18.0
tiktoken>=0.5.0

# Web requests
requests>=2.31.0
httpx>=0.24.0

# Configuration and environment
pydantic>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0

# Utilities (optional - comment out if not available)
# git+https://github.com/ClaimsBoost/claimsboost-utilities.git